Refactor last step
