+++ src/components/SetupProfile.vue
@@ -14,10 +14,7 @@ function nextStep() {
 <template>
   <div class="space-y-4">
     <SettingsProfileBlock context="setup" />
-
     <SettingsLinkBlock context="setup" />
-
-    <SettingsSubSpacesBlock context="setup" />
   </div>
   <div class="px-4 md:px-0">
     <SetupButtonBack @click="emit('back')" />
